pyproject.toml
[project]
name = "acme-proxy"
version = "0.1.0"
description = "ACME implementation in Python that proxies all local requests to actual ACME server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.5",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.7",
    "types-pygments>=2.19.0.20250715",
    "types-python-jose>=3.5.0.20250531",
    "types-pyyaml>=6.0.12.20250516",
    "types-ujson>=5.10.0.20250326",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
pretty = true
allow_redefinition = true
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = true
local_partial_types = true
no_implicit_optional = true
show_error_codes = true
strict = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = false
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true
